Split up the giant files LfpDisplayCanvas.cpp and LfpDisplayCanvas.h
