+++ Plugins/LfpDisplayNode/EventDisplayInterface.h
@@ -1,0 +1,66 @@
+/*
+    ------------------------------------------------------------------
+
+    This file is part of the Open Ephys GUI
+    Copyright (C) 2013 Open Ephys
+
+    ------------------------------------------------------------------
+
+    This program is free software: you can redistribute it and/or modify
+    it under the terms of the GNU General Public License as published by
+    the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    This program is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+*/
+#ifndef __EVENTDISPLAYINTERFACE_H__
+#define __EVENTDISPLAYINTERFACE_H__
+
+#include <VisualizerWindowHeaders.h>
+
+#include <vector>
+#include <array>
+
+#include "LfpDisplayClasses.h"
+#include "LfpDisplayNode.h"
+namespace LfpViewer {
+#pragma  mark - EventDisplayInterface -
+//==============================================================================
+/**
+    Interface class for Event Display channels.
+ */
+class EventDisplayInterface : public Component,
+    public Button::Listener
+{
+public:
+    EventDisplayInterface(LfpDisplay*, LfpDisplayCanvas*, int chNum);
+    ~EventDisplayInterface();
+
+    void paint(Graphics& g);
+
+    void buttonClicked(Button* button);
+
+    void checkEnabledState();
+
+    bool isEnabled;
+
+private:
+
+    int channelNumber;
+
+    LfpDisplay* display;
+    LfpDisplayCanvas* canvas;
+
+    ScopedPointer<UtilityButton> chButton;
+
+};
+    
+}; // namespace
+#endif
